SpeedDial fix aria-controls
"The aria-controls attribute identifies the element (or elements) whose contents or presence are controlled by the element on which the attribute is set" (--> https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-controls)

The aria-controls attribute should only be placed on the SpeedDial trigger button. And its value must be the identifier or list of identifiers of its items.